Bindings

Description

Documentation

idle :: IORef GLfloat -> IORef GLfloat -> IdleCallback

The idle function is called by GLUT every tick

display :: IORef GLfloat -> IORef (GLfloat, GLfloat) -> DisplayCallback

The display function is called by GLUT and at the moment displays a bunch of cubes in a circle

reshape :: ReshapeCallback

The reshape function is called whenever the window gets resized

keyboardMouse :: IORef GLfloat -> IORef (GLfloat, GLfloat) -> KeyboardMouseCallback

The keyboardMouse function is called whenever the user hits a key or does something with the mouse This one uses the keys:

space -> reverse direction

+ -> increase speed

- -> decrease speed

left -> move left

right -> move right

up -> move up

down -> move downs